requirements.txt
streamlit
PyPDF2
openai
python-dotenv
markdown2
xhtml2pdf
streamlit_option_menu